Reduce running time
The old solution was straightforward brute force, with a terrible worst
case running time – we might have to iterate through all n numbers n
times with a truly malicioius input. Now we instead add every element
to the summation set the first time we see it, and update the sum
accordingly. When the sum gets to large, we remove elements from the
head of the queue until this is no longer the case. This means
each element of the list is iterated over at most twice, and the
solution runs in linear time.